cleaned up frontend code
