+++ src/client/index.html
@@ -29,6 +29,6 @@
     <script>
       const global = globalThis;
     </script>
-    <script type="module" src="/src/Main.tsx"></script>
+    <script type="module" src="/src/index.tsx"></script>
   </body>
 </html>
