+++ src/client/src/App.tsx
@@ -1,35 +1,29 @@
 import { useContext } from "react";
-import { BrowserRouter, Navigate, Route, Routes } from "react-router-dom";
-import Home from "./Home";
-import Login from "./Login";
+import { Route, Routes } from "react-router-dom";
+import { ProtectedRoute } from "./hocs/ProtectedRoute";
+import Home from "./pages/Home";
+import Login from "./pages/Login";
 import { AuthContext } from "./providers/AuthProvider";
 
-import axios from "axios";
-import { useNavigate } from "react-router-dom";
-
 function App() {
-  const { isLoading, isAuthenticated } = useContext(AuthContext);
-  const navigate = useNavigate();
+  const { isLoading } = useContext(AuthContext);
 
-  axios.interceptors.response.use(
-    (response) => response,
-    (error) => {
-      if (error.response.status === 401) {
-        navigate("/login");
-      }
-      return Promise.reject(error);
-    }
-  );
+  if (isLoading) {
+    return <div>Loading...</div>;
+  }
+
   return (
-    <>
-      {isLoading && <div>Loading...</div>}
-      {!isLoading && (
-        <Routes>
-          <Route path="/" element={<Home />}></Route>
-          <Route path="/login" element={<Login />} />
-        </Routes>
-      )}
-    </>
+    <Routes>
+      <Route
+        path="/"
+        element={
+          <ProtectedRoute>
+            <Home />
+          </ProtectedRoute>
+        }
+      />
+      <Route path="/login" element={<Login />} />
+    </Routes>
   );
 }
